+++ libs/types/src/tranches.rs
@@ -12,6 +12,38 @@ pub enum TrancheType<Rate> {
 	},
 }
 
+impl<Rate> TrancheType<Rate>
+where
+	Rate: PartialOrd + PartialEq,
+{
+	/// Compares tranches with the following schema:
+	///
+	/// * (Residual, Residual) => true
+	/// * (Residual, NonResidual) => true,
+	/// * (NonResidual, Residual) => false,
+	/// * (NonResidual, NonResidual) =>
+	///         interest rate of next tranche must be smaller
+	///         equal to the interest rate of self.
+	///
+	pub fn valid_next_tranche(&self, next: &TrancheType<Rate>) -> bool {
+		match (self, next) {
+			(TrancheType::Residual, TrancheType::Residual) => true,
+			(TrancheType::Residual, TrancheType::NonResidual { .. }) => true,
+			(TrancheType::NonResidual { .. }, TrancheType::Residual) => false,
+			(
+				TrancheType::NonResidual {
+					interest_rate_per_sec: ref interest_prev,
+					..
+				},
+				TrancheType::NonResidual {
+					interest_rate_per_sec: ref interest_next,
+					..
+				},
+			) => interest_prev >= interest_next,
+		}
+	}
+}
+
 #[derive(Debug, Encode, PartialEq, Eq, Decode, Clone, TypeInfo)]
 pub struct TrancheMetadata<MaxTokenNameLength, MaxTokenSymbolLength>
 where
